requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via llama-index-core
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.21.0
    # via llama-index-core
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via llama-deploy
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via aiohttp
banks==2.1.2
    # via llama-index-core
beautifulsoup4==4.13.4
    # via llama-index-readers-file
brotli==1.1.0
    # via llama-deploy
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   llama-cloud
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   llama-cloud-services
    #   nltk
    #   uvicorn
colorama==0.4.6
    # via griffe
dataclasses-json==0.6.7
    # via llama-index-core
decorator==5.2.1
    # via ipython
deprecated==1.2.18
    # via
    #   banks
    #   llama-index-core
    #   llama-index-instrumentation
dirtyjson==1.0.8
    # via llama-index-core
distro==1.9.0
    # via openai
executing==2.2.0
    # via stack-data
fastapi==0.115.13
    # via llama-deploy
filetype==1.2.0
    # via llama-index-core
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via llama-index-core
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via llama-deploy
greenlet==3.2.3
    # via sqlalchemy
griffe==1.7.3
    # via banks
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   llama-cloud
    #   llama-index-core
    #   openai
    #   tavily-python
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipython==9.3.0
    # via pyvis
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   banks
    #   pyvis
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via nltk
jsonpickle==4.1.1
    # via pyvis
llama-cloud==0.1.26
    # via
    #   llama-cloud-services
    #   llama-index-indices-managed-llama-cloud
llama-cloud-services==0.6.34
    # via llama-parse
llama-deploy==0.8.1
    # via -r requirements.in
llama-index==0.12.43
    # via -r requirements.in
llama-index-agent-openai==0.4.11
    # via
    #   llama-index
    #   llama-index-program-openai
llama-index-cli==0.4.3
    # via llama-index
llama-index-core==0.12.43
    # via
    #   llama-cloud-services
    #   llama-deploy
    #   llama-index
    #   llama-index-agent-openai
    #   llama-index-cli
    #   llama-index-embeddings-openai
    #   llama-index-indices-managed-llama-cloud
    #   llama-index-llms-openai
    #   llama-index-multi-modal-llms-openai
    #   llama-index-program-openai
    #   llama-index-question-gen-openai
    #   llama-index-readers-file
    #   llama-index-readers-llama-parse
    #   llama-index-utils-workflow
llama-index-embeddings-openai==0.3.1
    # via
    #   llama-index
    #   llama-index-cli
llama-index-indices-managed-llama-cloud==0.7.7
    # via llama-index
llama-index-instrumentation==0.2.0
    # via llama-index-workflows
llama-index-llms-openai==0.4.7
    # via
    #   llama-index
    #   llama-index-agent-openai
    #   llama-index-cli
    #   llama-index-multi-modal-llms-openai
    #   llama-index-program-openai
    #   llama-index-question-gen-openai
llama-index-multi-modal-llms-openai==0.5.1
    # via llama-index
llama-index-program-openai==0.3.2
    # via
    #   llama-index
    #   llama-index-question-gen-openai
llama-index-question-gen-openai==0.3.1
    # via llama-index
llama-index-readers-file==0.4.9
    # via llama-index
llama-index-readers-llama-parse==0.4.0
    # via llama-index
llama-index-utils-workflow==0.3.5
    # via -r requirements.in
llama-index-workflows==0.2.2
    # via llama-index-core
llama-parse==0.6.34
    # via llama-index-readers-llama-parse
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
multidict==6.5.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via llama-index-core
networkx==3.5
    # via
    #   llama-index-core
    #   pyvis
nltk==3.9.1
    # via
    #   llama-index
    #   llama-index-core
numpy==2.3.0
    # via
    #   llama-index-core
    #   pandas
openai==1.88.0
    # via
    #   llama-index-agent-openai
    #   llama-index-embeddings-openai
    #   llama-index-llms-openai
packaging==25.0
    # via marshmallow
pandas==2.2.3
    # via llama-index-readers-file
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via llama-index-core
platformdirs==4.3.8
    # via
    #   banks
    #   llama-cloud-services
    #   llama-deploy
prometheus-client==0.21.1
    # via llama-deploy
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   banks
    #   fastapi
    #   llama-cloud
    #   llama-cloud-services
    #   llama-deploy
    #   llama-index-core
    #   llama-index-instrumentation
    #   llama-index-workflows
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via llama-deploy
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pypdf==5.6.0
    # via llama-index-readers-file
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   llama-cloud-services
    #   llama-deploy
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.18
    # via llama-deploy
pytz==2025.2
    # via pandas
pyvis==0.3.2
    # via llama-index-utils-workflow
pyyaml==6.0.2
    # via
    #   llama-deploy
    #   llama-index-core
    #   uvicorn
regex==2024.11.6
    # via
    #   nltk
    #   tiktoken
requests==2.32.4
    # via
    #   llama-index-core
    #   tavily-python
    #   tiktoken
rich==13.9.4
    # via llama-deploy
ruff==0.12.0
    # via -r requirements.in
setuptools==80.9.0
    # via llama-index-core
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.41
    # via llama-index-core
stack-data==0.6.3
    # via ipython
starlette==0.46.2
    # via fastapi
striprtf==0.0.26
    # via llama-index-readers-file
tavily-python==0.7.7
    # via -r requirements.in
tenacity==9.1.2
    # via llama-index-core
tiktoken==0.9.0
    # via
    #   llama-index-core
    #   tavily-python
tqdm==4.67.1
    # via
    #   llama-index-core
    #   nltk
    #   openai
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.14.0
    # via
    #   aiosqlite
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   llama-deploy
    #   llama-index-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.34.3
    # via llama-deploy
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via
    #   deprecated
    #   llama-index-core
yarl==1.20.1
    # via aiohttp
